tool/requirements.txt
pillow
pngquant
